+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `local_admin_unit` property to `observer`
 - `MOBILE_FORM_DELAYED_REALTIME` added to `source` enum
 - `IMPRECISE` added to `precision` enum
+- `project_name` added to `observation` model
 
 ### Changed